refactor(core): move shared file model out of app
