+++ src/core/mod.rs
@@ -1,0 +1,4 @@
+mod file_model;
+
+pub(crate) use self::file_model::FileClass;
+pub use self::file_model::{Entry, EntryKind, SortMode};
