+++ src/ui/theme/mod.rs
@@ -1,7 +1,7 @@
 mod appearance;
 mod builtin_themes;
 
-use crate::app::Entry;
+use crate::core::{Entry, EntryKind};
 use ratatui::style::Color;
 use std::path::Path;
 
@@ -37,18 +37,18 @@ pub(super) fn entry_symbol(entry: &Entry) -> &'static str {
 pub(super) fn path_color(path: &Path, is_dir: bool, palette: Palette) -> Color {
     let _ = palette;
     let kind = if is_dir {
-        crate::app::EntryKind::Directory
+        EntryKind::Directory
     } else {
-        crate::app::EntryKind::File
+        EntryKind::File
     };
     resolve_path(path, kind).color
 }
 
 pub(super) fn path_symbol(path: &Path, is_dir: bool) -> &'static str {
     let kind = if is_dir {
-        crate::app::EntryKind::Directory
+        EntryKind::Directory
     } else {
-        crate::app::EntryKind::File
+        EntryKind::File
     };
     resolve_path(path, kind).icon
 }
